Update writing.11tydata.js
You're getting an extra, unwanted HTML file (likely an empty or partially rendered article.html or index.html within an article folder) that corresponds directly to your article.njk template file, in addition to the correctly generated individual article pages.

This is a common scenario when a template file (like article.njk) is used for pagination, but Eleventy also tries to render it as a standalone page. Our current writing.11tydata.js is supposed to prevent this by setting permalink: false when data.post isn't defined, but it seems to not be working in this specific case.

Let's add some more explicit logging to your writing.11tydata.js to truly understand what the permalink function sees when it processes article.njk as a standalone file, and to ensure we're correctly accessing the slug from your Contentful data structure (a point of confusion earlier).